+++ CHANGELOG.md
@@ -179,6 +179,7 @@ It is also possible to install escripts and archives by providing a Git/GitHub r
 
 #### Elixir
 
+  * [Access] Do not accept nils in `Access.key/1` and `Access.key/2` in favor of explicit default values
   * [Float] Avoid multiple roundings in `Float.ceil/2`, `Float.floor/2` and `Float.round/2`
   * [Kernel] Don't crash in `macro_exported?/3` when dealing with Erlang modules
   * [Kernel] Ensure locals calls are rewritten when calling a local function or macro from inside a module
@@ -222,6 +223,7 @@ It is also possible to install escripts and archives by providing a Git/GitHub r
 
 #### Elixir
 
+  * [Access] `Access.key/1` is deprecated due to erratic behaviour for missing keys, please use Access.key/2 instead with proper default values
   * [Behaviour] The `Behaviour` module is deprecated. Callbacks may now be defined directly via the `@callback` attribute
   * [Enum] Deprecate `Enum.uniq/2` in favor of `Enum.uniq_by/2`
   * [Float] `Float.to_char_list/2` and `Float.to_string/2` are deprecated (use the :erlang functions if such conversions are desired)